add search bar component and function
