+++ .gitignore
@@ -30,3 +30,4 @@ coverage
 *.tsbuildinfo
 node_modules
 package-lock.json
+script.js
