+++ src/components/CanvasComponent.vue
@@ -1,0 +1,37 @@
+<template>
+
+  <button class="btn btn-primary" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasRight" aria-controls="offcanvasRight">Toggle right offcanvas</button>
+
+<div class="offcanvas offcanvas-end" tabindex="-1" id="offcanvasRight" aria-labelledby="offcanvasRightLabel">
+  <div class="offcanvas-header">
+    <h5 class="offcanvas-title" id="offcanvasRightLabel">Offcanvas right</h5>
+    <button type="button" class="btn-close" data-bs-dismiss="offcanvas" aria-label="Close"></button>
+  </div>
+  <div class="offcanvas-body">
+    ...
+  </div>
+</div>
+
+
+</template>
+
+<script>
+  import { store } from '../store';
+  export default {
+    name: 'CanvasComponent',
+    data() {
+      return {
+        store,
+        
+      }
+    },
+   
+
+
+  }
+</script>
+
+<style lang="scss" scoped>
+
+
+</style>
