+++ src/store.js
@@ -4,5 +4,8 @@ export const store = reactive({
 
     apiBaseUrl: "http://127.0.0.1:8000/api",
     imgBasePath: "http://127.0.0.1:8000/storage/",
+
+   
+    searchCity:'',
 });
